+++ code/tools/WtsTelemetry/WtsTelemetry/MailService.cs
@@ -1,0 +1,33 @@
+﻿using SendGrid.Helpers.Mail;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace WtsTelemetry
+{
+    public static class MailService
+    {
+        public static Mail CreateMail()
+        {
+            var personalization = new Personalization();
+            personalization.Tos = GetEmailReceivers().ToList();
+
+            var mail = new Mail();
+            mail.AddPersonalization(personalization);
+            mail.From = new Email(Environment.GetEnvironmentVariable("SendGrid:From"));
+            mail.Subject = Environment.GetEnvironmentVariable("SendGrid:Subject");
+            mail.AddContent(new Content("text/html", Environment.GetEnvironmentVariable("SendGrid:Content")));
+
+            return mail;
+        }
+
+        private static IEnumerable<Email> GetEmailReceivers()
+        {
+            var receivers = Environment.GetEnvironmentVariable("SendGrid:To").Split(';');
+            foreach (var receiver in receivers)
+            {
+                yield return new Email(receiver.Trim());
+            }
+        }
+    }
+}
